moved all hooks related to function toCourseSite() (including the function) out of <PrepPeriod /> and up into <CardHeader />, then passed toCourseSite() down to <Prep /> and <PrepPeriod />
