+++ src/components/Content/Card/CardHeader/PrepPeriod/PrepPeriod.js
@@ -1,22 +1,9 @@
-import React, { useContext } from "react";
-import { useNavigate } from 'react-router-dom';
-import PageContext from "../../../../../contexts/PageContext";
-import UserContext from "../../../../../contexts/userContext";
+import React from "react";
 import styles from './PrepPeriod.module.css';
 
-const PrepPeriod = ({ period, minimised }) => {
-    const { changePeriod } = useContext(UserContext);
-    const { changePage } = useContext(PageContext);
-    const navigate = useNavigate();
-
+const PrepPeriod = ({ period, minimised, toCourseSite }) => {
     const periodId = period.period === 'ct' ? 'people-group'
         : period.period === undefined ? '' : period.period[0];
-
-    const toCourseSite = e => {
-        changePage('Dashboard');
-        changePeriod(e);
-        navigate('../../../../../CourseSite');
-    }
 
     return (
         <i className={`
